updata to 190803

+++ plugins/convert/writer/ffmpeg.py
@@ -6,17 +6,20 @@ from math import ceil
 
 import imageio
 import imageio_ffmpeg as im_ffm
-from ffmpy import FFmpeg
+from ffmpy import FFmpeg, FFRuntimeError
 
 from ._base import Output, logger
 
 
 class Writer(Output):
     """ Video output writer using imageio """
-    def __init__(self, output_folder, total_count, source_video):
-        super().__init__(output_folder)
+    def __init__(self, output_folder, total_count, frame_ranges, source_video, **kwargs):
+        super().__init__(output_folder, **kwargs)
+        logger.debug("total_count: %s, frame_ranges: %s, source_video: '%s'",
+                     total_count, frame_ranges, source_video)
         self.source_video = source_video
-        self.frame_order = list(range(1, total_count + 1))
+        self.frame_ranges = frame_ranges
+        self.frame_order = self.set_frame_order(total_count)
         self.output_dimensions = None  # Fix dims of 1st frame in case of different sized images
         self.writer = None  # Need to know dimensions of first frame, so set writer then
 
@@ -50,6 +53,7 @@ class Writer(Output):
         """ Return the fps of source video """
         reader = imageio.get_reader(self.source_video)
         retval = reader.get_meta_data()["fps"]
+        reader.close()
         logger.debug(retval)
         return retval
 
@@ -77,6 +81,17 @@ class Writer(Output):
         logger.debug(output_args)
         return output_args
 
+    def set_frame_order(self, total_count):
+        """ Return the full list of frames to be converted in order """
+        if self.frame_ranges is None:
+            retval = list(range(1, total_count + 1))
+        else:
+            retval = list()
+            for rng in self.frame_ranges:
+                retval.extend(list(range(rng[0], rng[1] + 1)))
+        logger.debug("frame_order: %s", retval)
+        return retval
+
     def get_writer(self):
         """ Add the requested encoding options and return the writer """
         logger.debug("writer config: %s", self.config)
@@ -84,6 +99,7 @@ class Writer(Output):
                                   fps=self.video_fps,
                                   ffmpeg_log_level="error",
                                   quality=None,
+                                  macro_block_size=8,
                                   output_params=self.output_params)
 
     def write(self, filename, image):
@@ -130,6 +146,16 @@ class Writer(Output):
             however muxing audio is non-trivial, so this is done afterwards with ffmpy.
             A future fix could be implemented to mux audio with the frames """
         logger.info("Muxing Audio...")
+        if self.frame_ranges is not None:
+            logger.warning("Muxing audio is not currently supported for limited frame ranges."
+                           "The output video has been created but you will need to mux audio "
+                           "yourself")
+            os.rename(self.video_tmp_file, self.video_file)
+            logger.debug("Removing temp file")
+            if os.path.isfile(self.video_tmp_file):
+                os.remove(self.video_tmp_file)
+            return
+
         exe = im_ffm.get_ffmpeg_exe()
         inputs = OrderedDict([(self.video_tmp_file, None), (self.source_video, None)])
         outputs = {self.video_file: "-map 0:0 -map 1:1 -c: copy"}
@@ -138,6 +164,22 @@ class Writer(Output):
                      inputs=inputs,
                      outputs=outputs)
         logger.debug("Executing: %s", ffm.cmd)
-        ffm.run()
+        # Sometimes ffmpy exits for no discernible reason, but then works on a later attempt,
+        # so take 5 shots at this
+        attempts = 5
+        for attempt in range(attempts):
+            logger.debug("Muxing attempt: %s", attempt + 1)
+            try:
+                ffm.run()
+            except FFRuntimeError as err:
+                logger.debug("ffmpy runtime error: %s", str(err))
+                if attempt != attempts - 1:
+                    continue
+                logger.error("There was a problem muxing audio. The output video has been "
+                             "created but you will need to mux audio yourself either with the "
+                             "EFFMpeg tool or an external application.")
+                os.rename(self.video_tmp_file, self.video_file)
+            break
         logger.debug("Removing temp file")
-        os.remove(self.video_tmp_file)
+        if os.path.isfile(self.video_tmp_file):
+            os.remove(self.video_tmp_file)
